requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
Pillow==10.0.0
pytest==7.4.0
python-dateutil==2.8.2
flask-cors==5.0.1
websocket-client==1.6.3

# 使用国内镜像源安装依赖:
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple 